requirements.txt
pandas==1.5.3
numpy==1.23.5
streamlit==1.24.0
Pillow==9.5.0
scikit-learn==1.2.2
xgboost==1.7.6
catboost==1.1.1
matplotlib==3.7.2
seaborn==0.12.2